move composio service to workflows integrations directory
